reorganized testing directory
