+++ server/tests/helper/clearTestDb.js
